burger-constructor reorganized
